[Chore] #252 - dateformatter 함수화
