+++ iOS-NOTTODO/Widget-NOTTODO/Provider/TimeEntity.swift
@@ -11,16 +11,8 @@ struct SimpleEntry: TimelineEntry {
     var date: Date = .now
     
     var dayOfWeek: String {
-        return dateFormatterString(format: "E", date: date)
+        return Formatter.dateFormatterString(format: "E", date: date)
     }
     var todayMission: [DailyMissionResponseDTO]
     let quote: String
-    
-    private func dateFormatterString(format: String?, date: Date) -> String {
-        let formatter = Foundation.DateFormatter()
-        formatter.dateFormat = format ?? "yyyy-MM-dd"
-        formatter.locale = Locale(identifier: "ko_KR")
-        let convertStr = formatter.string(from: date)
-        return convertStr
-    }
 }
